서버에서 레벨 관리, 스폰된 아이템 삭제버그 수정, 넷플레이어 아이템 레벨, 죽템 스폰안되는 버그 수정

+++ Project_Animagus/Server/Server/Player.h
@@ -32,7 +32,7 @@ public:
     bool    isAlive = true;
     std::string name;
 
-    std::atomic<int8> powerUpLevel = 0;
+    std::atomic<int32> powerUpLevel = 0;
     
     Protocol::PlayerType type = Protocol::PlayerType::KOALA;     // TODO : 선택한 캐릭터로 보여지게
